refactor: move audit registration into AuditRegistry
Breakout from #607.
