+++ crates/zizmor/src/main.rs
@@ -9,7 +9,6 @@ use std::{
 use annotate_snippets::{Level, Renderer};
 use anstream::{eprintln, println, stream::IsTerminal};
 use anyhow::{Context, Result, anyhow};
-use audit::{Audit, AuditLoadError};
 use camino::{Utf8Path, Utf8PathBuf};
 use clap::{CommandFactory, Parser, ValueEnum};
 use clap_complete::Generator;
@@ -620,51 +619,7 @@ fn run() -> Result<ExitCode> {
         &audit_state,
     )?;
 
-    let mut audit_registry = AuditRegistry::new();
-    macro_rules! register_audit {
-        ($rule:path) => {{
-            // HACK: https://github.com/rust-lang/rust/issues/48067
-            use crate::audit::AuditCore as _;
-            use $rule as base;
-            match base::new(&audit_state) {
-                Ok(audit) => audit_registry.register_audit(base::ident(), Box::new(audit)),
-                Err(AuditLoadError::Skip(e)) => {
-                    tracing::info!("skipping {audit}: {e}", audit = base::ident())
-                }
-                Err(AuditLoadError::Fail(e)) => {
-                    return Err(anyhow!(tips(
-                        format!("failed to load audit: {audit}", audit = base::ident()),
-                        &[format!("{e:#}"), format!("see: {url}", url = base::url())]
-                    )));
-                }
-            }
-        }};
-    }
-
-    register_audit!(audit::artipacked::Artipacked);
-    register_audit!(audit::unsound_contains::UnsoundContains);
-    register_audit!(audit::excessive_permissions::ExcessivePermissions);
-    register_audit!(audit::dangerous_triggers::DangerousTriggers);
-    register_audit!(audit::impostor_commit::ImpostorCommit);
-    register_audit!(audit::ref_confusion::RefConfusion);
-    register_audit!(audit::use_trusted_publishing::UseTrustedPublishing);
-    register_audit!(audit::template_injection::TemplateInjection);
-    register_audit!(audit::hardcoded_container_credentials::HardcodedContainerCredentials);
-    register_audit!(audit::self_hosted_runner::SelfHostedRunner);
-    register_audit!(audit::known_vulnerable_actions::KnownVulnerableActions);
-    register_audit!(audit::unpinned_uses::UnpinnedUses);
-    register_audit!(audit::insecure_commands::InsecureCommands);
-    register_audit!(audit::github_env::GitHubEnv);
-    register_audit!(audit::cache_poisoning::CachePoisoning);
-    register_audit!(audit::secrets_inherit::SecretsInherit);
-    register_audit!(audit::bot_conditions::BotConditions);
-    register_audit!(audit::overprovisioned_secrets::OverprovisionedSecrets);
-    register_audit!(audit::unredacted_secrets::UnredactedSecrets);
-    register_audit!(audit::forbidden_uses::ForbiddenUses);
-    register_audit!(audit::obfuscation::Obfuscation);
-    register_audit!(audit::stale_action_refs::StaleActionRefs);
-    register_audit!(audit::unpinned_images::UnpinnedImages);
-    register_audit!(audit::anonymous_definition::AnonymousDefinition);
+    let audit_registry = AuditRegistry::default_audits(&audit_state);
 
     let mut results = FindingRegistry::new(&app, &config);
     {
