+++ crates/zizmor/src/registry.rs
@@ -16,10 +16,11 @@ use thiserror::Error;
 
 use crate::{
     App,
-    audit::{Audit, AuditInput},
+    audit::{self, Audit, AuditInput},
     config::Config,
     finding::{Confidence, Finding, Persona, Severity},
     models::{action::Action, workflow::Workflow},
+    state::AuditState,
 };
 
 #[derive(Error, Debug)]
@@ -264,10 +265,57 @@ pub(crate) struct AuditRegistry {
 }
 
 impl AuditRegistry {
-    pub(crate) fn new() -> Self {
+    fn empty() -> Self {
         Self {
             audits: Default::default(),
         }
+    }
+
+    /// Constructs a new [`AuditRegistry`] with all default audits registered.
+    pub(crate) fn default_audits(audit_state: &AuditState) -> Self {
+        let mut registry = Self::empty();
+
+        macro_rules! register_audit {
+            ($rule:path) => {{
+                // HACK: https://github.com/rust-lang/rust/issues/48067
+                use $rule as base;
+
+                use crate::audit::AuditCore as _;
+                match base::new(&audit_state) {
+                    Ok(audit) => registry.register_audit(base::ident(), Box::new(audit)),
+                    Err(e) => tracing::info!(
+                        target: "zizmor", "skipping {audit}: {e}", audit = base::ident()
+                    ),
+                }
+            }};
+        }
+
+        register_audit!(audit::artipacked::Artipacked);
+        register_audit!(audit::unsound_contains::UnsoundContains);
+        register_audit!(audit::excessive_permissions::ExcessivePermissions);
+        register_audit!(audit::dangerous_triggers::DangerousTriggers);
+        register_audit!(audit::impostor_commit::ImpostorCommit);
+        register_audit!(audit::ref_confusion::RefConfusion);
+        register_audit!(audit::use_trusted_publishing::UseTrustedPublishing);
+        register_audit!(audit::template_injection::TemplateInjection);
+        register_audit!(audit::hardcoded_container_credentials::HardcodedContainerCredentials);
+        register_audit!(audit::self_hosted_runner::SelfHostedRunner);
+        register_audit!(audit::known_vulnerable_actions::KnownVulnerableActions);
+        register_audit!(audit::unpinned_uses::UnpinnedUses);
+        register_audit!(audit::insecure_commands::InsecureCommands);
+        register_audit!(audit::github_env::GitHubEnv);
+        register_audit!(audit::cache_poisoning::CachePoisoning);
+        register_audit!(audit::secrets_inherit::SecretsInherit);
+        register_audit!(audit::bot_conditions::BotConditions);
+        register_audit!(audit::overprovisioned_secrets::OverprovisionedSecrets);
+        register_audit!(audit::unredacted_secrets::UnredactedSecrets);
+        register_audit!(audit::forbidden_uses::ForbiddenUses);
+        register_audit!(audit::obfuscation::Obfuscation);
+        register_audit!(audit::stale_action_refs::StaleActionRefs);
+        register_audit!(audit::unpinned_images::UnpinnedImages);
+        register_audit!(audit::anonymous_definition::AnonymousDefinition);
+
+        registry
     }
 
     pub(crate) fn len(&self) -> usize {
